requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
google-cloud-storage